Fix a bunch of compiler warnings. Added Release64 and Debug64 build targets in VS
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@16622 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ qt/simcpaperdoll.cc
@@ -681,7 +681,7 @@ ItemDataListModel::ItemDataListModel( QObject* parent ) :
 int
 ItemDataListModel::rowCount( const QModelIndex& /* parent */ ) const
 {
-  return dbc::n_items( false );
+  return static_cast<int> ( dbc::n_items( false ) );
 }
 
 QVariant
@@ -1290,7 +1290,7 @@ EnchantDataModel::EnchantDataModel( PaperdollProfile* profile, QObject* parent )
 int
 EnchantDataModel::rowCount( const QModelIndex& ) const
 {
-  return m_enchants.size();
+  return static_cast<int>( m_enchants.size() );
 }
 
 QVariant
